docs/_tooling/requirements.txt
--extra-index-url https://pypi.org/simple/

Sphinx

# At least 4.2.0, as it fixes a bug in combination with esbonio live preview:
# https://github.com/useblocks/sphinx-needs/issues/1350
sphinx-needs>=4.2.0

sphinxcontrib-plantuml
pydata-sphinx-theme
sphinx-design
sphinxcontrib-plantuml
sphinx-autobuild
ruamel.yaml

# Until release of esbonio 1.x, we need to install it ourselves so the VS Code esbonio-extension
# can find it.
# esbonio >= 1 comes bundled with the esbonio-extension >= 1.
# esbonio<1 is required for the esbonio-extension <1
esbonio<1

# Although not required in all targets, we want pytest within ide_support to run tests from the IDE.
pytest
pytest-cov
